Moves IsTotal[x] methods to respective derived classes. Returns 0 total miles if the last row is not a total miles entry.
