+++ Appserver/FormSubmit/MileageForm.cs
@@ -1,6 +1,9 @@
 ﻿using Newtonsoft.Json;
 using System.Collections.Generic;
 using System.Linq;
+using System;
+using Appserver.TextractDocument;
+using System.Text.RegularExpressions;
 
 namespace Appserver.FormSubmit
 {
@@ -27,8 +30,7 @@ namespace Appserver.FormSubmit
         public void addMileRow(string date, string miles, string group, string purpose) =>
             this.Mileage.Add(new MileageRowItem(date, miles, group, purpose));
 
-
-        protected override void AddTables(List<TextractDocument.Table> tables)
+        protected override void AddTables(List<Table> tables)
         {
             var table = tables[0].GetTable();
             // Remove first row
@@ -66,6 +68,58 @@ namespace Appserver.FormSubmit
             {
                 totalMiles = lastrow[1].ToString().Trim();
             }
+            else
+            {
+                totalMiles = "0";
+            }
+        }
+
+
+        public static bool isTotalMilesRow(List<Cell> lastrow)
+        {
+            Regex rxNumberGroup = new Regex(@"([0-9])+");
+            int confidence = 0;
+            bool foundDate = false;
+
+            foreach (var entry in lastrow)
+            {
+                var s = entry.ToString().Trim().ToLower();
+
+                // If we can parse the entry into a valid
+                // DateTime format, probably areen't dealing
+                // with a total miles row.
+                try
+                {
+                    var date = DateTime.Parse(s).ToString("yyyy-MM-dd");
+                    foundDate = true;
+                }
+                catch (FormatException) { }
+
+                // Increase confidence we have a total miles row
+                if (s.Contains("total"))
+                {
+                    confidence++;
+                }
+
+                if (s.Contains("miles"))
+                {
+                    confidence++;
+                }
+
+                var matches = rxNumberGroup.Matches(s);
+                confidence += matches.Count;
+            }
+
+            // Most likely a mileage entry
+            if (foundDate)
+            {
+                return false;
+            }
+
+            // Did we at least find a contains words
+            // and at least one group of numbers that could
+            // be a mileage total?
+            return confidence >= 2;
         }
     }
 }
