+++ Appserver/FormSubmit/TimesheetForm.cs
@@ -3,6 +3,7 @@ using System.Collections.Generic;
 using Appserver.TextractDocument;
 using System;
 using System.Linq;
+using System.Text.RegularExpressions;
 
 public class TimesheetForm: AbstractFormObject
 {
@@ -80,5 +81,33 @@ public class TimesheetForm: AbstractFormObject
                 totalHours = lastrow[3].ToString();
             }
         }
+    }
+
+    // If one of the rows in the cell matches the regex
+    // and there is also a row that can be parsed into hours,
+    // then we probably have a total hours cell. 
+    public static bool isTotalTimeRow(List<Cell> lastrow)
+    {
+        // The letters of "total/unit/hours" without
+        // "yes/no" for group field.
+        Regex rxTotal = new Regex(@"([uthrli])+");
+        Regex rxTime = new Regex(@"([0-9])+");
+        bool matchRex = false;
+        bool matchTime = false;
+
+        foreach (var entry in lastrow)
+        {
+            // TotalHours match?
+            var s = entry.ToString().Trim().ToLower();
+            var totalmatches = rxTotal.Matches(s);
+            if (totalmatches.Count >= 1) { matchRex = true; }
+
+            // Time match?
+            var time = FixHours(entry.ToString()).ToString().Trim();
+            var timeMatches = rxTime.Matches(time.ToLower());
+            if (timeMatches.Count >= 1) { matchTime = true; }
+        }
+
+        return matchRex && matchTime;
     }
 }
